requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
markdown-it-py==2.2.0
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.0
Pillow==10.0.0
proglog==0.1.10
Pygments==2.15.1
PyJWT==2.6.0
PySocks==1.7.1
pytz==2023.3
requests==2.29.0
requests-toolbelt==1.0.0
rich==13.3.5
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
